Improved performance on 'harder' tasks
Tasks that are not just about adding a card or deck.
Content-based search currently is broken; could not be tested!

The remaining, harder interaction tests now work *kindof* well, meaning that in general the llm either passes or it is on the right track, making spelling mistakes, or selecting a wrong deck with similar name etc.